2L0K Persistent HAMT (#15)
HAMTs are made persistent; use hamt_with to create a new HAMT with a new
key/value pair from a previous version. This is used for scopes by
keeping a stack of scopes and adding to the previous one, then popping
it from the stack and freeing it when leaving the scope. (Unless nothing
was added to it).